requirements.txt
amqp==2.6.1
appdirs==1.4.4
asgiref==3.2.10
attrs==20.3.0
Automat==20.2.0
autopep8==1.4.4
Babel==2.9.0
billiard==3.6.3.0
bootstrap4==0.1.0
braintree==3.59.0
cairocffi==1.2.0
CairoSVG==2.5.0
celery==4.4.2
certifi==2020.12.5
cffi==1.14.3
chardet==3.0.4
constantly==15.1.0
cryptography==3.1.1
cssselect2==0.4.1
defusedxml==0.6.0
distlib==0.3.1
Django==3.0.11
django-allauth==0.39.1
django-braces==1.14.0
django-cors-headers==3.5.0
django-countries==5.3.3
django-crispy-forms==1.10.0
django-debug-toolbar==1.10.1
django-embed-video==1.3.2
django-extensions==2.2.5
django-localflavor==3.0.1
django-memcache-status==2.2
django-parler==2.0.1
django-rosetta==0.9.3
django-social-share==2.1.0
django-taggit==1.2.0
djangorestframework==3.11.0
easy-thumbnails==2.7
filelock==3.0.12
flower==0.9.3
html5lib==1.1
hyperlink==20.0.1
idna==2.8
incremental==17.5.0
kombu==4.6.11
Markdown==3.2.1
oauthlib==3.0.1
pep8==1.7.1
Pillow==7.0.0
pipenv==2020.8.13
polib==1.1.0
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.0.0
Pyphen==0.10.0
python-decouple==3.1
python-memcached==1.59
python-stdnum==1.14
python3-openid==3.1.0
pytz==2018.5
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.2.0
six==1.15.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.2.4
stripe==2.27.0
tinycss2==1.1.0
toml==0.10.1
tornado==5.1.1
txaio==20.4.1
urllib3==1.24.2
vine==1.3.0
virtualenv==20.0.31
virtualenv-clone==0.5.4
WeasyPrint==51
webencodings==0.5.1
Werkzeug==0.16.0
zope.interface==5.2.0